res fixed a bug in bindoc where requesting a float value when it was   stored as an int could return bogus results.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@19158 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/documentsystem/binary/bindoc.cpp
@@ -231,7 +231,7 @@ const char* csBinaryDocAttribute::GetValue ()
 	{
   	  char buf[50];
 	  cs_snprintf (buf, sizeof (buf) - 1, "%d", 
-	    (int)little_endian_long (attrPtr->value));
+	    (int)((int32)little_endian_long (attrPtr->value)));
 	  delete[] vstr; 
 	  vstr = csStrNew (buf);
 	  vsptr = attrPtr;
@@ -295,7 +295,7 @@ float csBinaryDocAttribute::GetValueAsFloat ()
       }
     case BD_VALUE_TYPE_INT:
       {
-	return (float)((int)little_endian_long (attrPtr->value));
+	return (float)((int32)little_endian_long (attrPtr->value));
       }
     case BD_VALUE_TYPE_FLOAT:
       {
